Modified Files:  	config.cc client/Buffercache.cc client/Buffercache.h  	client/Client.cc include/config.h
Much more stable but still some bugs.
Client::flush_buffers is currently blocking until buffer cache puts state changes
in critical regions.


git-svn-id: https://ceph.svn.sf.net/svnroot/ceph@440 29311d96-e01e-0410-9327-a35deaab8ce9

+++ ceph/config.h
@@ -40,8 +40,8 @@ struct md_config_t {
   int      client_bcache_alloc_maxsize;
   int      client_bcache_ttl;
   int      client_bcache_size;
-  float    client_bcache_lowater;
-  float    client_bcache_hiwater;
+  int      client_bcache_lowater;
+  int      client_bcache_hiwater;
   int      client_bcache_maxfrag;
 
   int      client_trace;
